Set the type to uint32 to not loose precision between the original 16bits integer PNGs and the resulting OpenEXR file

+++ apps/tools/cave-exr/main.cpp
@@ -180,7 +180,7 @@ int main(int argc, char* argv[])
         allowedPixelTypes.push_back("uint32");
 
         TCLAP::ValuesConstraint<std::string> allowedPixelTypesVals(allowedPixelTypes);
-        TCLAP::ValueArg<std::string> pixelTypeArg("t", "type", "Pixel type to use in the resulting OpenEXR file.", false, "float", &allowedPixelTypesVals);
+        TCLAP::ValueArg<std::string> pixelTypeArg("t", "type", "Pixel type to use in the resulting OpenEXR file.", false, "uint32", &allowedPixelTypesVals);
 
         cmd.add(pixelTypeArg);
